+++ sdk/src/trace/samplers/parent_factory.cc
@@ -1,0 +1,24 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+#include "opentelemetry/sdk/trace/samplers/parent_factory.h"
+#include "opentelemetry/sdk/trace/samplers/parent.h"
+
+namespace trace_api = opentelemetry::trace;
+
+OPENTELEMETRY_BEGIN_NAMESPACE
+namespace sdk
+{
+namespace trace
+{
+
+std::unique_ptr<Sampler> ParentBasedSamplerFactory::Create(
+    std::shared_ptr<Sampler> delegate_sampler)
+{
+  std::unique_ptr<Sampler> sampler(new ParentBasedSampler(delegate_sampler));
+  return sampler;
+}
+
+}  // namespace trace
+}  // namespace sdk
+OPENTELEMETRY_END_NAMESPACE